Integrated the World class with the Game code
Removed input from the game. We will add this in a more structured
way soon.
Had a number of bugs that I needed to fix.
1. Forgot #pragma once at the top of Texture.hpp
2. SceneNode::Ptr instead of Ptr
3. #include RenderWindow.hpp in World.cpp
4. Game.hpp. It is crucial to define the window before the world in
this class. If not the world tries to work with a window which has
not been created yet and you get messed up window, camera and spawm
positions