Adding explosions to the game
When an aircraft is destroyed it now explodes. Explosions are handled
with the Animation class which steps through a sprite sheet to display
images in sequence and make it look like an explosion. Need to be
careful about removing entities after the explosion is finished and
only allowing exploded planes possibly drop one pickup