Everything looks better with Shaders
Now we use a RenderTexture rather than always using a RenderWindow.
We can apply the Shader to the RenderTexture before rendering in
the RenderWindow.
When you stare at a bright light for too long your vision
goes kind of blurry with spots. Don't try it!
This is the effect we implement in this code. Note the shaders
are defined and written in GLSL and are in the Media/Shaders folder
To understand them you will need to look at a GLSL tutorial -
http://www.lighthouse3d.com/tutorials/glsl-tutorial/