Everything sounds better with music
Added music to the menu and game. When MenuState and GameState are
constructed the music starts playing.When in PauseState the
music is paused and when back to game the music resumes.
Note SFML does not support mp3 format so need to use something like
Audacity to export files as .ogg
Note that you need to add sfml-audio-d.lib to Properties->Linker->
Additonal Dependencies otherwise you will get a bunch of linker errors